RH-SSO testcases sync
