--- security/keycloak-multitenant/src/test/java/io/quarkus/ts/security/keycloak/multitenant/OpenShiftRhSso75MultiTenantSecurityIT.java
@@ -1,34 +1,0 @@
-package io.quarkus.ts.security.keycloak.multitenant;
-
-import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
-
-import io.quarkus.test.bootstrap.KeycloakService;
-import io.quarkus.test.bootstrap.RestService;
-import io.quarkus.test.scenarios.OpenShiftScenario;
-import io.quarkus.test.services.Container;
-import io.quarkus.test.services.QuarkusApplication;
-
-@OpenShiftScenario
-@EnabledIfSystemProperty(named = "ts.redhat.registry.enabled", matches = "true")
-public class OpenShiftRhSso75MultiTenantSecurityIT extends BaseMultiTenantSecurityIT {
-
-    static final int KEYCLOAK_PORT = 8080;
-
-    @Container(image = "${rhsso.75.image}", expectedLog = "Http management interface listening", port = KEYCLOAK_PORT)
-    static KeycloakService keycloak = new KeycloakService(REALM_DEFAULT)
-            .withProperty("SSO_IMPORT_FILE", "resource::/keycloak-realm.json");
-
-    @QuarkusApplication
-    static RestService app = new RestService()
-            .withProperty("quarkus.oidc.auth-server-url", () -> keycloak.getRealmUrl());
-
-    @Override
-    protected KeycloakService getKeycloak() {
-        return keycloak;
-    }
-
-    @Override
-    protected RestService getApp() {
-        return app;
-    }
-}
