+++ security/keycloak-oauth2/src/test/java/io/quarkus/ts/security/keycloak/oauth2/OpenShiftRhSsoOauth2SecurityIT.java
@@ -1,4 +1,4 @@
-package io.quarkus.ts.security.keycloak.authz;
+package io.quarkus.ts.security.keycloak.oauth2;
 
 import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
 
@@ -10,19 +10,18 @@ import io.quarkus.test.services.QuarkusApplication;
 
 @OpenShiftScenario
 @EnabledIfSystemProperty(named = "ts.redhat.registry.enabled", matches = "true")
-public class OpenShiftRhSso75AuthzSecurityIT extends BaseAuthzSecurityIT {
+public class OpenShiftRhSsoOauth2SecurityIT extends BaseOauth2SecurityIT {
 
     static final int KEYCLOAK_PORT = 8080;
 
-    @Container(image = "${rhsso.75.image}", expectedLog = "Http management interface listening", port = KEYCLOAK_PORT)
+    @Container(image = "${rhsso.image}", expectedLog = "Http management interface listening", port = KEYCLOAK_PORT)
     static KeycloakService keycloak = new KeycloakService(REALM_DEFAULT)
             .withProperty("SSO_IMPORT_FILE", "resource::/keycloak-realm.json");
 
     @QuarkusApplication
     static RestService app = new RestService()
-            .withProperty("quarkus.oidc.auth-server-url", () -> keycloak.getRealmUrl())
-            .withProperty("quarkus.oidc.client-id", CLIENT_ID_DEFAULT)
-            .withProperty("quarkus.oidc.credentials.secret", CLIENT_SECRET_DEFAULT);
+            .withProperty("quarkus.oauth2.introspection-url",
+                    () -> keycloak.getRealmUrl() + "/protocol/openid-connect/token/introspect");
 
     @Override
     protected KeycloakService getKeycloak() {
