+++ security/keycloak-oidc-client-basic/src/test/java/io/quarkus/ts/security/keycloak/oidcclient/basic/OpenShiftRhSsoOidcClientSecurityIT.java
@@ -1,4 +1,6 @@
-package io.quarkus.ts.security.keycloak.oauth2;
+package io.quarkus.ts.security.keycloak.oidcclient.basic;
+
+import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
 
 import io.quarkus.test.bootstrap.KeycloakService;
 import io.quarkus.test.bootstrap.RestService;
@@ -7,18 +9,18 @@ import io.quarkus.test.services.Container;
 import io.quarkus.test.services.QuarkusApplication;
 
 @OpenShiftScenario
-public class OpenShiftRhSso73Oauth2SecurityIT extends BaseOauth2SecurityIT {
+@EnabledIfSystemProperty(named = "ts.redhat.registry.enabled", matches = "true")
+public class OpenShiftRhSsoOidcClientSecurityIT extends BaseOidcClientSecurityIT {
 
     static final int KEYCLOAK_PORT = 8080;
 
-    @Container(image = "${rhsso.73.image}", expectedLog = "Http management interface listening", port = KEYCLOAK_PORT)
+    @Container(image = "${rhsso.image}", expectedLog = "Http management interface listening", port = KEYCLOAK_PORT)
     static KeycloakService keycloak = new KeycloakService(REALM_DEFAULT)
             .withProperty("SSO_IMPORT_FILE", "resource::/keycloak-realm.json");
 
     @QuarkusApplication
     static RestService app = new RestService()
-            .withProperty("quarkus.oauth2.introspection-url",
-                    () -> keycloak.getRealmUrl() + "/protocol/openid-connect/token/introspect");
+            .withProperty("quarkus.oidc.auth-server-url", () -> keycloak.getRealmUrl());
 
     @Override
     protected KeycloakService getKeycloak() {
