--- security/keycloak-webapp/src/test/java/io/quarkus/ts/security/keycloak/webapp/OpenShiftRhSso73WebappSecurityIT.java
@@ -1,32 +1,0 @@
-package io.quarkus.ts.security.keycloak.webapp;
-
-import io.quarkus.test.bootstrap.KeycloakService;
-import io.quarkus.test.bootstrap.RestService;
-import io.quarkus.test.scenarios.OpenShiftScenario;
-import io.quarkus.test.services.Container;
-import io.quarkus.test.services.QuarkusApplication;
-
-@OpenShiftScenario
-public class OpenShiftRhSso73WebappSecurityIT extends BaseWebappSecurityIT {
-
-    static final int KEYCLOAK_PORT = 8080;
-
-    @Container(image = "${rhsso.73.image}", expectedLog = "Http management interface listening", port = KEYCLOAK_PORT)
-    static KeycloakService keycloak = new KeycloakService(REALM_DEFAULT)
-            .withProperty("SSO_IMPORT_FILE", "resource::/keycloak-realm.json");
-
-    @QuarkusApplication
-    static RestService app = new RestService()
-            .withProperty("quarkus.oidc.auth-server-url", () -> keycloak.getRealmUrl())
-            .withProperty("quarkus.oidc.client-id", CLIENT_ID_DEFAULT);
-
-    @Override
-    protected KeycloakService getKeycloak() {
-        return keycloak;
-    }
-
-    @Override
-    protected RestService getApp() {
-        return app;
-    }
-}
