+++ security/oidc-client-mutual-tls/src/test/java/io/quarkus/ts/security/oidcclient/mtls/OpenShiftRhSsoOidcMtlsIT.java
@@ -2,6 +2,8 @@ package io.quarkus.ts.security.oidcclient.mtls;
 
 import static io.quarkus.ts.security.oidcclient.mtls.MutualTlsKeycloakService.newRhSsoInstance;
 
+import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
+
 import io.quarkus.test.bootstrap.KeycloakService;
 import io.quarkus.test.bootstrap.RestService;
 import io.quarkus.test.scenarios.OpenShiftScenario;
@@ -9,20 +11,21 @@ import io.quarkus.test.services.Container;
 import io.quarkus.test.services.QuarkusApplication;
 
 @OpenShiftScenario
-public class OpenShiftRhSso73OidcMtlsIT extends KeycloakMtlsAuthN {
+@EnabledIfSystemProperty(named = "ts.redhat.registry.enabled", matches = "true")
+public class OpenShiftRhSsoOidcMtlsIT extends KeycloakMtlsAuthN {
 
-    @Container(image = "${rhsso.73.image}", expectedLog = EXPECTED_LOG, port = KEYCLOAK_PORT)
-    static KeycloakService rhsso = newRhSsoInstance();
+    @Container(image = "${rhsso.image}", expectedLog = EXPECTED_LOG, port = KEYCLOAK_PORT)
+    static KeycloakService rhsso = newRhSsoInstance().withRedHatFipsDisabled();
 
     /**
      * Keystore file type is automatically detected by file extension by quarkus-oidc.
      */
     @QuarkusApplication
-    static RestService app = createRestService(PKCS12_KEY_STORE_FILE_EXTENSION, "", rhsso::getRealmUrl);
+    static RestService app = createRestService(JKS_KEY_STORE_FILE_EXTENSION, "", rhsso::getRealmUrl);
 
     @Override
     protected String getKeyStoreFileExtension() {
-        return PKCS12_KEY_STORE_FILE_EXTENSION;
+        return JKS_KEY_STORE_FILE_EXTENSION;
     }
 
     @Override
